Move hymod_params struct to dedicated file.
